+++ Assets/Scripts/Game/Puzzle/Light/LightZone.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 87334ab5885137142a65d9ddc25f7915
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
