--- Assets/Scripts/Game/Puzzle/Light/PerformanceSaver.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 59ca839a2ba252141bbea3cf208e08e2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
