+++ Assets/Scripts/Game/Puzzle/Light/Prefabs/LightZone.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2d6ba5032f44d94dbd1c49b5eff93f1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
