+++ ProjectSettings/TagManager.asset
@@ -32,7 +32,7 @@ TagManager:
   - Player
   - Platform
   - Obstacles
-  - 
+  - LightZone
   - 
   - 
   -
